Move CSV parsing functions into transformation.py
